+++ src/builder/styles/index.ts
@@ -1,0 +1,2 @@
+export * from './monochrome.class';
+export * from './text.class';