+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './direction.enum';
 export * from './flags.interface';
 export * from './format.enum';
+export * from './skinparams.interface';