+++ src/types/skinparams.interface.ts
@@ -1,0 +1,9 @@
+import { Direction } from './direction.enum';
+
+export interface SkinParams {
+	direction?: Direction,
+	roundcorner?: number,
+	linetype?: 'ortho' | 'polyline',
+	shadowing?: 'true' | 'false',
+	handwritten?: 'true' | 'false',
+}