+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased] - TBD
 
 - Updated dependencies to the latest versions.
+- Updated UML builder to handle styles in a separate class.
 
 ## [v1.5.0] (2020-11-28)